Renamed device_functions.h/cpp -> device_test_functions.h/cpp
Was blocking travis due to CUDA deprecation.
